Split home redesign visual sections
